Update taken space on file upload

+++ Models/User.cs
@@ -16,7 +16,7 @@ public class User
     [BsonElement("phone_no")] public string? PhoneNo { get; set; }
     [BsonElement("lang")] public Language? Language = Models.Language.EN;
     [BsonElement("icon")] public string? Icon { get; set; } = "";
-    [BsonElement("taken_space")] public float? TakenSpace { get; set; } = 0;
+    [BsonElement("taken_space")] public long? TakenSpace { get; set; } = 0;
     [BsonElement("docs")] public List<Document> Docs { get; set; } = new List<Document>();
     [BsonElement("categories")] public List<string> Categories { get; set; } = new List<string>();
     [BsonElement("medical_visits")] public List<MedicalVisit> MedicalVisits { get; set; } = new List<MedicalVisit>();
